requirement.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
langchain-community
sqlalchemy
psycopg2-binary
requests
python-dotenv
pydantic